Refactors output file writers out of IchnosCF
